separate group.py and group_utils.py
